+++ src/main/java/hu/gaborneorcsity/fruits/processors/attribute_extractors/FruitSizeExtractor.java
@@ -1,5 +1,13 @@
 package hu.gaborneorcsity.fruits.processors.attribute_extractors;
 
+/**
+ * A service for extracting the size of an HTML document containing a single fruit
+ */
 public interface FruitSizeExtractor {
-    double extract(String document);
+    /**
+     * Extracts the size of the provided HTML document containing a single fruit
+     * @param document the document whose size is extracted
+     * @return the extracted size in bytes
+     */
+    int extract(String document);
 }
