+++ src/main/java/hu/gaborneorcsity/fruits/processors/attribute_extractors/JSoupFruitSizeExtractor.java
@@ -2,11 +2,17 @@ package hu.gaborneorcsity.fruits.processors.attribute_extractors;
 
 import java.io.UnsupportedEncodingException;
 
+/**
+ * A simple implementation based on JSoup
+ */
 public class JSoupFruitSizeExtractor implements FruitSizeExtractor {
+    /**
+     * {@inheritDoc}
+     */
     @Override
-    public double extract(String document) {
+    public int extract(String document) {
         try {
-            return (double) document.getBytes("UTF8").length;
+            return document.getBytes("UTF8").length;
         } catch (UnsupportedEncodingException e) {
             throw new RuntimeException(e);
         }
